Finished most error handling

+++ src/sample/DatabaseConnection/RemoveAlbumDatabase.java
@@ -6,7 +6,8 @@ import javafx.scene.control.ButtonType;
 import java.sql.*;
 
 //Class that removes an album from the db
-public class RemoveAlbumDatabase {
+public class RemoveAlbumDatabase
+{
 
     private Statement st;
     private Connection connection;
@@ -17,7 +18,8 @@ public class RemoveAlbumDatabase {
     private String songKey;
 
     //Constructor that connects to the db
-    public RemoveAlbumDatabase(){
+    public RemoveAlbumDatabase()
+    {
         try
         {
             connection = DriverManager.getConnection(url, username, password);
@@ -31,19 +33,54 @@ public class RemoveAlbumDatabase {
         }
     }
 
+
     //Method to remove the album with the given id
-    public void removeAlbum(int id) {
+    public void removeAlbum(int id)
+    {
+        try
+        {
+            //Query that selects id album
+            String searchQuery = "SELECT idalbum FROM album WHERE idalbum = " + id + ";";
 
-        String query = "DELETE FROM album WHERE idalbum = ?";
-        try (PreparedStatement ps = connection.prepareStatement(query)) {
-            ps.setInt(1, id);
+            st = connection.createStatement();
+            ResultSet rs = st.executeQuery(searchQuery);
 
-            ps.executeUpdate();
-        } catch (SQLException e) {
-            e.printStackTrace();
-            Alert alert = new Alert(Alert.AlertType.ERROR, "Something went wrong trying to save the data in the Database", ButtonType.OK);
-            alert.showAndWait();
+            //Check if id exists
+            if(rs.next())
+            {
+                //Remove album
+                String query = "DELETE FROM album WHERE idalbum = ?";
+                try (PreparedStatement ps = connection.prepareStatement(query))
+                {
+                    ps.setInt(1, id);
+                    ps.executeUpdate();
+
+                    //Message for complete
+                    Alert alert = new Alert(Alert.AlertType.INFORMATION, "Successfully removed the selected album from the database.", ButtonType.OK);
+                    alert.setHeaderText("REMOVAL COMPLETE");
+                    alert.showAndWait();
+                }
+                catch (SQLException e)
+                {
+                    Alert alert = new Alert(Alert.AlertType.ERROR, "Something went wrong trying to delete the album from the database.", ButtonType.OK);
+                    alert.setHeaderText("ERROR");
+                    alert.showAndWait();
+                }
+            }
+            else
+            {
+                //Id does not exist message
+                Alert alert = new Alert(Alert.AlertType.ERROR, "The album id you filled in does not exist.", ButtonType.OK);
+                alert.setHeaderText("ERROR");
+                alert.showAndWait();
+            }
         }
+        catch (SQLException e)
+        {
+
+        }
+
+
     }
